display: use qti_kernel_headers for installclean build time reduction
Old KERNEL_OBJ method of including kernel headers is really bad on
incremental builds because it causes the targets to regenerate every
time a new build is made. Switch to new qti_kernel_headers header lib to
fix this problem.

Time delta with this is about 1.5min on my machine (from total build
time of ~7min).

Test: builds, boots to home screen

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>
Change-Id: Ie4c946787839562c1eb3ee33c790bbae3e284443